Reply function and exception handling

+++ src/main/java/cn/orekiyuta/ark/service/QuestionService.java
@@ -76,7 +76,7 @@ public class QuestionService {
         return paginationDTO;
     }
 
-    public PaginationDTO list(Integer userId, Integer page, Integer size) {
+    public PaginationDTO list(Long userId, Integer page, Integer size) {
 
         PaginationDTO paginationDTO = new PaginationDTO();
 
@@ -126,7 +126,7 @@ public class QuestionService {
         return paginationDTO;
     }
 
-    public QuestionDTO getById(Integer id) {
+    public QuestionDTO getById(Long id) {
         Question question=questionMapper.selectByPrimaryKey(id);
         if (question == null){
             throw new CustomizeException(CustomizeErrorCode.QUESTION_NOT_FOUND);
@@ -141,6 +141,7 @@ public class QuestionService {
     public void createOrUpdate(Question question) {
 
         if(question.getId() == null){
+            //创建
             question.setGmtCreate(System.currentTimeMillis());
             question.setGmtModified(question.getGmtCreate());
 //            question.setViewCount(0);
@@ -148,6 +149,7 @@ public class QuestionService {
 //            question.setCommentCount(0);
             questionMapper.insertSelective(question);
         }else {
+            //更新
 //            question.setGmtModified(question.getGmtCreate());
             Question updateQuestion = new Question();
             updateQuestion.setGmtModified(System.currentTimeMillis());
@@ -165,7 +167,7 @@ public class QuestionService {
 
     }
 
-    public void incView(Integer id) {
+    public void incView(Long id) {
 
         Question question = new Question();
         question.setId(id);
